requirements.txt
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
django-mathfilters==1.0.0
Python version == 3.7.9 
spacy==3.0.3
pytz==2021.1
